+++ test/test_debug_pp.expected.ml
@@ -23,19 +23,25 @@ let bar (x : t) =
                  ((();
                    Debug_runtime.log_value_pp ?descr:(Some "y")
                      ~entry_id:__entry_id ~pp:pp_num ~is_result:true y);
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                    ~start_lnum:8 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                    ~start_lnum:8 ~entry_id:__entry_id;
+                  raise e)) in
           x.second * y
     with
     | __res ->
         (Debug_runtime.log_value_pp ?descr:(Some "bar") ~entry_id:__entry_id
            ~pp:pp_num ~is_result:true __res;
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:7
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  num)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:7
+           ~entry_id:__entry_id;
+         raise e)) : num)
 let () = ignore @@ (bar { first = 7; second = 42 })
 let baz (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -56,19 +62,25 @@ let baz (x : t) =
                  ((();
                    Debug_runtime.log_value_pp ?descr:(Some "_yz")
                      ~entry_id:__entry_id ~pp ~is_result:true _yz);
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                    ~start_lnum:14 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                    ~start_lnum:14 ~entry_id:__entry_id;
+                  raise e)) in
           (x.second * y) + z
     with
     | __res ->
         (Debug_runtime.log_value_pp ?descr:(Some "baz") ~entry_id:__entry_id
            ~pp:pp_num ~is_result:true __res;
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:13
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  num)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:13
+           ~entry_id:__entry_id;
+         raise e)) : num)
 let () = ignore @@ (baz { first = 7; second = 42 })
 let rec loop (depth : num) (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -101,10 +113,13 @@ let rec loop (depth : num) (x : t) =
                       ((();
                         Debug_runtime.log_value_pp ?descr:(Some "y")
                           ~entry_id:__entry_id ~pp:pp_num ~is_result:true y);
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                         ~start_lnum:23 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                         ~start_lnum:23 ~entry_id:__entry_id;
+                       raise e)) in
                let z : num =
                  let __entry_id = Debug_runtime.get_entry_id () in
                  ();
@@ -118,17 +133,23 @@ let rec loop (depth : num) (x : t) =
                       ((();
                         Debug_runtime.log_value_pp ?descr:(Some "z")
                           ~entry_id:__entry_id ~pp:pp_num ~is_result:true z);
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                         ~start_lnum:24 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_pp.ml"
+                         ~start_lnum:24 ~entry_id:__entry_id;
+                       raise e)) in
                z + 7)
     with
     | __res ->
         (Debug_runtime.log_value_pp ?descr:(Some "loop") ~entry_id:__entry_id
            ~pp:pp_num ~is_result:true __res;
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:19
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  num)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_pp.ml" ~start_lnum:19
+           ~entry_id:__entry_id;
+         raise e)) : num)
 let () = ignore @@ (loop 0 { first = 7; second = 42 })