+++ test/test_debug_sexp.expected.ml
@@ -21,20 +21,26 @@ let foo (x : int) =
                    Debug_runtime.log_value_sexp ?descr:(Some "y")
                      ~entry_id:__entry_id ~is_result:true
                      (([%sexp_of : int]) y));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:8 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:8 ~entry_id:__entry_id;
+                  raise e)) in
           [x; y; 2 * y]
     with
     | __res ->
         (Debug_runtime.log_value_sexp ?descr:(Some "foo")
            ~entry_id:__entry_id ~is_result:true
            (([%sexp_of : int list]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:7
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int list)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:7
+           ~entry_id:__entry_id;
+         raise e)) : int list)
 let () = ignore @@ (List.hd @@ (foo 7))
 type t = {
   first: int ;
@@ -59,19 +65,25 @@ let bar (x : t) =
                    Debug_runtime.log_value_sexp ?descr:(Some "y")
                      ~entry_id:__entry_id ~is_result:true
                      (([%sexp_of : int]) y));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:16 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:16 ~entry_id:__entry_id;
+                  raise e)) in
           x.second * y
     with
     | __res ->
         (Debug_runtime.log_value_sexp ?descr:(Some "bar")
            ~entry_id:__entry_id ~is_result:true (([%sexp_of : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:15
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:15
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (bar { first = 7; second = 42 })
 let baz (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -93,10 +105,13 @@ let baz (x : t) =
                    Debug_runtime.log_value_sexp ?descr:(Some "_yz")
                      ~entry_id:__entry_id ~is_result:true
                      (([%sexp_of : (int * int)]) _yz));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:22 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:22 ~entry_id:__entry_id;
+                  raise e)) in
           let (((u, w) as _uw) : (int * int)) =
             let __entry_id = Debug_runtime.get_entry_id () in
             ();
@@ -109,19 +124,25 @@ let baz (x : t) =
                    Debug_runtime.log_value_sexp ?descr:(Some "_uw")
                      ~entry_id:__entry_id ~is_result:true
                      (([%sexp_of : (int * int)]) _uw));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:23 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:23 ~entry_id:__entry_id;
+                  raise e)) in
           (((x.second * y) + z) + u) + w
     with
     | __res ->
         (Debug_runtime.log_value_sexp ?descr:(Some "baz")
            ~entry_id:__entry_id ~is_result:true (([%sexp_of : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:21
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:21
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (baz { first = 7; second = 42 })
 let lab ~x:(x : int)  =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -143,20 +164,26 @@ let lab ~x:(x : int)  =
                    Debug_runtime.log_value_sexp ?descr:(Some "y")
                      ~entry_id:__entry_id ~is_result:true
                      (([%sexp_of : int]) y));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:29 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                    ~start_lnum:29 ~entry_id:__entry_id;
+                  raise e)) in
           [x; y; 2 * y]
     with
     | __res ->
         (Debug_runtime.log_value_sexp ?descr:(Some "lab")
            ~entry_id:__entry_id ~is_result:true
            (([%sexp_of : int list]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:28
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int list)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:28
+           ~entry_id:__entry_id;
+         raise e)) : int list)
 let () = ignore @@ (List.hd @@ (lab ~x:7))
 let rec loop (depth : int) (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -190,10 +217,13 @@ let rec loop (depth : int) (x : t) =
                         Debug_runtime.log_value_sexp ?descr:(Some "y")
                           ~entry_id:__entry_id ~is_result:true
                           (([%sexp_of : int]) y));
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                         ~start_lnum:38 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                         ~start_lnum:38 ~entry_id:__entry_id;
+                       raise e)) in
                let z : int =
                  let __entry_id = Debug_runtime.get_entry_id () in
                  ();
@@ -208,17 +238,23 @@ let rec loop (depth : int) (x : t) =
                         Debug_runtime.log_value_sexp ?descr:(Some "z")
                           ~entry_id:__entry_id ~is_result:true
                           (([%sexp_of : int]) z));
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                         ~start_lnum:39 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_sexp.ml"
+                         ~start_lnum:39 ~entry_id:__entry_id;
+                       raise e)) in
                z + 7)
     with
     | __res ->
         (Debug_runtime.log_value_sexp ?descr:(Some "loop")
            ~entry_id:__entry_id ~is_result:true (([%sexp_of : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:34
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_sexp.ml" ~start_lnum:34
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (loop 0 { first = 7; second = 42 })