+++ test/test_debug_show.expected.ml
@@ -19,19 +19,25 @@ let foo (x : int) =
                  ((();
                    Debug_runtime.log_value_show ?descr:(Some "y")
                      ~entry_id:__entry_id ~is_result:true (([%show : int]) y));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:5 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:5 ~entry_id:__entry_id;
+                  raise e)) in
           [x; y; 2 * y]
     with
     | __res ->
         (Debug_runtime.log_value_show ?descr:(Some "foo")
            ~entry_id:__entry_id ~is_result:true (([%show : int list]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:4
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int list)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:4
+           ~entry_id:__entry_id;
+         raise e)) : int list)
 let () = ignore @@ (List.hd @@ (foo 7))
 type t = {
   first: int ;
@@ -55,19 +61,25 @@ let bar (x : t) =
                  ((();
                    Debug_runtime.log_value_show ?descr:(Some "y")
                      ~entry_id:__entry_id ~is_result:true (([%show : int]) y));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:13 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:13 ~entry_id:__entry_id;
+                  raise e)) in
           x.second * y
     with
     | __res ->
         (Debug_runtime.log_value_show ?descr:(Some "bar")
            ~entry_id:__entry_id ~is_result:true (([%show : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:12
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:12
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (bar { first = 7; second = 42 })
 let baz (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -89,19 +101,25 @@ let baz (x : t) =
                    Debug_runtime.log_value_show ?descr:(Some "_yz")
                      ~entry_id:__entry_id ~is_result:true
                      (([%show : (int * int)]) _yz));
-                  Debug_runtime.close_log ~entry_id:__entry_id;
+                  Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:19 ~entry_id:__entry_id;
                   __res)
              | exception e ->
-                 (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                 (Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                    ~start_lnum:19 ~entry_id:__entry_id;
+                  raise e)) in
           (x.second * y) + z
     with
     | __res ->
         (Debug_runtime.log_value_show ?descr:(Some "baz")
            ~entry_id:__entry_id ~is_result:true (([%show : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:18
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:18
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (baz { first = 7; second = 42 })
 let rec loop (depth : int) (x : t) =
   (let __entry_id = Debug_runtime.get_entry_id () in
@@ -135,10 +153,13 @@ let rec loop (depth : int) (x : t) =
                         Debug_runtime.log_value_show ?descr:(Some "y")
                           ~entry_id:__entry_id ~is_result:true
                           (([%show : int]) y));
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                         ~start_lnum:28 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                         ~start_lnum:28 ~entry_id:__entry_id;
+                       raise e)) in
                let z : int =
                  let __entry_id = Debug_runtime.get_entry_id () in
                  ();
@@ -153,17 +174,23 @@ let rec loop (depth : int) (x : t) =
                         Debug_runtime.log_value_show ?descr:(Some "z")
                           ~entry_id:__entry_id ~is_result:true
                           (([%show : int]) z));
-                       Debug_runtime.close_log ~entry_id:__entry_id;
+                       Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                         ~start_lnum:29 ~entry_id:__entry_id;
                        __res)
                   | exception e ->
-                      (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) in
+                      (Debug_runtime.close_log ~fname:"test_debug_show.ml"
+                         ~start_lnum:29 ~entry_id:__entry_id;
+                       raise e)) in
                z + 7)
     with
     | __res ->
         (Debug_runtime.log_value_show ?descr:(Some "loop")
            ~entry_id:__entry_id ~is_result:true (([%show : int]) __res);
-         Debug_runtime.close_log ~entry_id:__entry_id;
+         Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:24
+           ~entry_id:__entry_id;
          __res)
-    | exception e -> (Debug_runtime.close_log ~entry_id:__entry_id; raise e)) : 
-  int)
+    | exception e ->
+        (Debug_runtime.close_log ~fname:"test_debug_show.ml" ~start_lnum:24
+           ~entry_id:__entry_id;
+         raise e)) : int)
 let () = ignore @@ (loop 0 { first = 7; second = 42 })